Adding js pagination to community notes widget
YV-336 #comment This comes with some caveats. First, the API uses cursors instead of pages for paginating search results. Thus a "state" is needed to persist the cursor to advance to the next result set. This is not a problem when you have app clients b/c there is state between api requests, but without adding a Memcache key, there isn't on web. We *could* add a js array with a long list of cursors, but each cursor is a 256 char string. And if someone was crazy enough to paginate very far, it could cause a browser crash. A cursor would have to be stored for each page returned b/c the API does not support a direction passed to its search api. So you can't just use one cursor, you have to store a list of cursors. Anyways, I've removed the "Previous" pagination (similar to how users search works right now on web).
YV-336 #close