Support for API global strings
The API will sometimes use "str" instead of "l_str" if we have a global
string that does not need localized. The "str" field can be used
directly and does not need pulled from the PO file.
That being said, I'm not sure if this is enough to fix the issue, but I
wanted to throw it out there so Matt or Russ can make it better.