Adjust Logic that Determines When to Show "Start Plan" Button
*When displaying a Plan Started or Plan Completed moment:*
 * If I am the user that generated the moment (I started or completed
the plan) then the "Start Plan" button will never show
 * If a friend of mine generated the moment the "Start Plan" button
will only show if I am not already subscribed to the plan

*Possible Issue:*
 * If I recently completed a plan (I am no longer subscribed) and my
friend starts or completes the same plan, when their moment shows up on
my timeline, it will have a "Start Plan" button (even though I recently
completed it). As far as I know, there's no way to determine if I was
recently subscribed to a plan in the past.