Fixing all my Probs Yo
Technically, the respond_to failed me. Some clients are not sending the
proper header and HTML is attempting to render instead of JSON, and
there’s an error in the HTML template resulting from a missing var that
I removed while cleaning up. FIX: Just render JSON and forget HTML for
now.