Be Explicit About What We Send to t()
This fixes a XSS vulnerability. By default, HAML escapes HTML strings.
However, we are passing the string through t() and t is sending back a
translation not found tag that encloses the injected HTML. This fix
requires the source param to be one of the 7 options in the list that
we actually have translations for or we don’t display it.